Update filter functions in domUpdates
